requirements.txt
Django
pywebview